Exception handling - delete

+++ src/main/java/com/bernardo/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.bernardo.course.entities.User;
 import com.bernardo.course.repositories.UserRepository;
+import com.bernardo.course.services.exceptions.DatabaseException;
 import com.bernardo.course.services.exceptions.ResourceNotFoundeException;
 
 @Service
@@ -29,8 +31,16 @@ public class UserService {
 		return repository.save(obj);
 	}
 
-	public void delete(long id) {
-		repository.deleteById(id);
+	public void delete(Long id) {
+		try {
+			if (!repository.existsById(id))
+				throw new ResourceNotFoundeException(id);
+			repository.deleteById(id);
+		} catch (ResourceNotFoundeException e) {
+			throw new ResourceNotFoundeException(id);
+		} catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 
 	public User update(long id, User obj) {
